# docs/v1/swagger.yaml
basePath: /v1
definitions:
...
  protocol.DataType:
    enum:
    - object
    - number
    - integer
    - string
    - array
    - "null"
    - boolean
    type: string
    x-enum-varnames:
    - ObjectT
    - Number
    - Integer
    - String
    - Array
    - "Null"
    - Boolean
  protocol.InputSchema:
    properties:
      properties:
        additionalProperties:
          $ref: '#/definitions/protocol.Property'
        type: object
      required:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/protocol.InputSchemaType'
    type: object
  protocol.InputSchemaType:
    enum:
    - object
    type: string
    x-enum-varnames:
    - Object
  protocol.OutputSchema:
    properties:
      properties:
        additionalProperties:
          $ref: '#/definitions/protocol.Property'
        type: object
      required:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/protocol.InputSchemaType'
    type: object
  protocol.Property:
    properties:
      description:
        description: Description is the description of the schema.
        type: string
      enum:
        items:
          type: string
        type: array
      items:
        allOf:
        - $ref: '#/definitions/protocol.Property'
        description: Items specifies which data type an array contains, if the schema
          type is Array.
      properties:
        additionalProperties:
          $ref: '#/definitions/protocol.Property'
        description: Properties describes the properties of an object, if the schema
          type is Object.
        type: object
      required:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/protocol.DataType'
    type: object
  protocol.Tool:
    properties:
      annotations:
        allOf:
        - $ref: '#/definitions/protocol.ToolAnnotations'
        description: Annotations provides additional hints about the tool's behavior
      description:
        description: Description is a human-readable description of the tool
        type: string
      inputSchema:
        allOf:
        - $ref: '#/definitions/protocol.InputSchema'
        description: InputSchema defines the expected parameters for the tool using
          JSON Schema
      name:
        description: Name is the unique identifier of the tool
        type: string
      outputSchema:
        allOf:
        - $ref: '#/definitions/protocol.OutputSchema'
        description: OutputSchema defines expected output structure for the tool using
          Optional JSON Schema
    type: object
  protocol.ToolAnnotations:
    properties:
      destructiveHint:
        description: |-
          DestructiveHint indicates the tool may perform destructive updates
          (only meaningful when ReadOnlyHint is false)
        type: boolean
      idempotentHint:
        description: |-
          IdempotentHint indicates calling the tool repeatedly with the same arguments
          has no additional effect (only meaningful when ReadOnlyHint is false)
        type: boolean
      openWorldHint:
        description: OpenWorldHint indicates the tool may interact with an "open world"
          of external entities
        type: boolean
      readOnlyHint:
        description: ReadOnlyHint indicates the tool does not modify its environment
        type: boolean
      title:
        description: Title is a human-readable title for the tool, useful for UI display
        type: string
    type: object
...
  request.Avatar:
    properties:
      key:
        description: 前端透传给后端用于保存avatar，例如：custom-upload/avatar/abc/def.png
        type: string
      path:
        description: 前端请求地址，例如：/v1/static/avatar/abc/def.png (请求非必填)
        type: string
    type: object
...
  request.MCPCreate:
    properties:
      desc:
        description: 描述
        type: string
      from:
        description: 来源
        type: string
      mcpSquareId:
        description: 广场mcpId(非空表示来源于广场)
        type: string
      name:
        description: 名称
        type: string
      sseUrl:
        description: SSE URL
        type: string
    required:
    - desc
    - from
    - name
    - sseUrl
    type: object
  request.MCPIDReq:
    properties:
      mcpId:
        type: string
    required:
    - mcpId
    type: object
...
  request.MCPUpdate:
    properties:
      desc:
        description: 描述
        type: string
      from:
        description: 来源
        type: string
      mcpId:
        type: string
      name:
        description: 名称
        type: string
      sseUrl:
        description: SSE URL
        type: string
    required:
    - desc
    - from
    - mcpId
    - name
    - sseUrl
    type: object
...
  response.ListResult:
    properties:
      list: {}
      total:
        type: integer
    type: object
...
  response.MCPActionList:
    properties:
      actions:
        items:
          $ref: '#/definitions/protocol.Tool'
        type: array
    type: object
  response.MCPDetail:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      category:
        description: 类型(data:数据,create:创作,search:搜索)
        type: string
      desc:
        description: 描述
        type: string
      detail:
        description: 详情
        type: string
      feature:
        description: 特性说明
        type: string
      from:
        description: 来源
        type: string
      manual:
        description: 使用说明
        type: string
      mcpId:
        description: mcpId
        type: string
      mcpSquareId:
        description: 广场mcpId(非空表示来源于广场)
        type: string
      name:
        description: 名称
        type: string
      scenario:
        description: 应用场景
        type: string
      sseUrl:
        description: SSE URL
        type: string
      summary:
        description: 使用概述
        type: string
    type: object
  response.MCPInfo:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      category:
        description: 类型(data:数据,create:创作,search:搜索)
        type: string
      desc:
        description: 描述
        type: string
      from:
        description: 来源
        type: string
      mcpId:
        description: mcpId
        type: string
      mcpSquareId:
        description: 广场mcpId(非空表示来源于广场)
        type: string
      name:
        description: 名称
        type: string
      sseUrl:
        description: SSE URL
        type: string
    type: object
...
  response.MCPSelect:
    properties:
      description:
        description: 描述
        type: string
      mcpId:
        description: mcpId
        type: string
      mcpSquareId:
        description: 广场mcpId(非空表示来源于广场)
        type: string
      name:
        description: 名称
        type: string
      serverFrom:
        description: 来源
        type: string
      serverUrl:
        description: sseUrl
        type: string
      toolId:
        description: 工具id
        type: string
      toolName:
        description: 工具名称
        type: string
      toolType:
        description: 工具类型
        enum:
        - mcp
        - mcpserver
        type: string
      type:
        type: string
      uniqueId:
        description: 随机unique id(每次动态生成)
        type: string
    required:
    - toolType
    type: object
...
  response.MCPSquareDetail:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      category:
        description: 类型(data:数据,create:创作,search:搜索)
        type: string
      desc:
        description: 描述
        type: string
      detail:
        description: 详情
        type: string
      feature:
        description: 特性说明
        type: string
      from:
        description: 来源
        type: string
      hasCustom:
        description: 是否已经发送到自定义
        type: boolean
      manual:
        description: 使用说明
        type: string
      mcpSquareId:
        description: 广场mcpId(非空表示来源于广场)
        type: string
      name:
        description: 名称
        type: string
      scenario:
        description: 应用场景
        type: string
      sseUrl:
        description: SSE URL
        type: string
      summary:
        description: 使用概述
        type: string
      tools:
        description: 工具列表
        items:
          $ref: '#/definitions/protocol.Tool'
        type: array
    type: object
  response.MCPSquareInfo:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      category:
        description: 类型(data:数据,create:创作,search:搜索)
        type: string
      desc:
        description: 描述
        type: string
      from:
        description: 来源
        type: string
      mcpSquareId:
        description: 广场mcpId(非空表示来源于广场)
        type: string
      name:
        description: 名称
        type: string
    type: object
  response.MCPToolList:
    properties:
      tools:
        items:
          $ref: '#/definitions/protocol.Tool'
        type: array
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: "## HTTP Header\r\n| Header        | 说明      |\r\n| ------------- |
    --------- |\r\n| Authorization | JWT token |\r\n| X-Language    | 语言Code  |\r\n|
    X-Org-Id      | 组织ID    |\r\n| X-Client-Id   | 客户端标识|\r\n\r\n## HTTP Status\r\n|
    HTTP Status             | 说明                   |\r\n| -----------------------
    | ---------------------- |\r\n| 200, StatusOK           | 请求返回成功           |\r\n|
    400, StatusBadRequest   | 请求返回失败，用于业务 |\r\n| 401, StatusUnauthorized | JWT认证失败
    \           |\r\n| 403, StatusForbidden    | 没有权限               |\r\n\r\n## 权限-菜单对应表\r\n|
    一级权限        | 二级权限  | 三级权限 | 一级菜单 | 二级菜单 | 三级菜单 |\r\n|-------------|-------|------|------|------|------|\r\n|
    guest       |       |      | 【访客】 |      |      |\r\n| common      |       |      |
    【通用】 |      |      |\r\n| permission  |       |      | 权限管理 |      |      |\r\n|
    permission  | user  |      | 权限管理 | 用户管理 |      |\r\n| permission  | org   |      |
    权限管理 | 组织管理 |      |\r\n| permission  | role  |      | 权限管理 | 角色管理 |      |\r\n\r\n##
    `/v1/user/permission`返回用例\r\n```json\r\n{\r\n  \"code\": 0,\r\n  \"data\": {\r\n
    \   \"orgPermission\": {\r\n      \"org\": {\"id\": \"test-org-id\", \"name\":
    \"test-org-name\"},\r\n      \"permissions\": [\r\n        {\"perm\": \"permission\"},\r\n
    \       {\"perm\": \"permission.user\"},\r\n        {\"perm\": \"permission.org\"},\r\n
...
    \"操作成功\"\r\n}\r\n```"
...
  version: v0.0.1
paths:
...
  /mcp:
    delete:
      consumes:
      - application/json
      description: 删除自定义MCP
      parameters:
      - description: mcpId
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.MCPIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除自定义MCP
      tags:
      - tool
    get:
      consumes:
      - application/json
      description: 获取自定义MCP详情
      parameters:
      - description: mcpId
        in: query
        name: mcpId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.MCPDetail'
              type: object
      summary: 获取自定义MCP详情
      tags:
      - tool
    post:
      consumes:
      - application/json
      description: 创建自定义MCP
      parameters:
      - description: 自定义MCP信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.MCPCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建自定义MCP
      tags:
      - tool
    put:
      consumes:
      - application/json
      description: 修改自定义MCP
      parameters:
      - description: 自定义MCP信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.MCPUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 修改自定义MCP
      tags:
      - tool
  /mcp/action/list:
    get:
      consumes:
      - application/json
      description: 获取MCP Action列表
      parameters:
      - in: query
        name: toolId
        required: true
        type: string
      - enum:
        - mcp
        - mcpserver
        in: query
        name: toolType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.MCPActionList'
              type: object
      summary: 获取MCP Action列表
      tags:
      - mcp
  /mcp/list:
    get:
      consumes:
      - application/json
      description: 获取自定义MCP列表
      parameters:
      - description: mcp名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.MCPInfo'
                        type: array
                    type: object
              type: object
      summary: 获取自定义MCP列表
      tags:
      - tool
  /mcp/select:
    get:
      consumes:
      - application/json
      description: 获取自定义MCP列表（用于下拉选择）
      parameters:
      - description: mcp名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.MCPSelect'
                        type: array
                    type: object
              type: object
      summary: 获取自定义MCP列表（用于下拉选择）
      tags:
      - mcp
...
  /mcp/square:
    get:
      consumes:
      - application/json
      description: 获取广场MCP详情
      parameters:
      - description: mcpSquareId
        in: query
        name: mcpSquareId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.MCPSquareDetail'
              type: object
      summary: 获取广场MCP详情
      tags:
      - mcp.square
  /mcp/square/list:
    get:
      consumes:
      - application/json
      description: 获取广场MCP列表
      parameters:
      - description: mcp类型
        enum:
        - all
        - data
        - create
        - search
        in: query
        name: category
        type: string
      - description: mcp名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.MCPSquareInfo'
                        type: array
                    type: object
              type: object
      summary: 获取广场MCP列表
      tags:
      - mcp.square
  /mcp/square/recommend:
    get:
      consumes:
      - application/json
      description: 获取广场MCP推荐列表
      parameters:
      - description: mcpId
        in: query
        name: mcpId
        type: string
      - description: mcpSquareId
        in: query
        name: mcpSquareId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.MCPSquareInfo'
                        type: array
                    type: object
              type: object
      summary: 获取广场MCP推荐列表
      tags:
      - mcp.square
  /mcp/tool/list:
    get:
      consumes:
      - application/json
      description: 获取MCP Tool列表
      parameters:
      - description: mcpId(和serverUrl传一个)
        in: query
        name: mcpId
        type: string
      - description: serverUrl,就是sseUrl(和mcpId传一个)
        in: query
        name: serverUrl
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.MCPToolList'
              type: object
      summary: 获取MCP Tool列表
      tags:
      - mcp
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"